refactor(agent): split mcp_proxy_only.py into a src/ package
Address review feedback that the single 2800-line module was unwieldy. Split
it into focused modules under agent/src/, keeping @app.route handlers, client
code, and utilities separate; behavior is unchanged (bodies moved verbatim).

- src/config.py — config load, GCS/CONFIG_URL bootstrap, Secret Manager keys
- src/session_logger.py — audit logger + token-usage accumulator
- src/mcp/ — JSON-RPC client + session state, Gemini schema transform, data utils
- src/gemini/ — REST client (key rotation, streaming), JSON schemas
- src/workflows/ — MCP tool loop, KB search, chart config, follow-ups
- src/analytics/ — log parsing
- src/server/ — Flask app + routes; main.py is the entrypoint

The MCP `session_id`/`tools_cache` module globals stay in src/mcp/client.py and
are read via module-qualified access so mutations propagate. Config/logs are
anchored to AGENT_ROOT so they resolve to agent/ (same on-disk paths as before).
Dockerfile now COPYs main.py + src/ and runs main.py.
