+++ narratives/agent/main.py
@@ -1,0 +1,67 @@
+#!/usr/bin/env python3
+# Copyright 2024 Google LLC
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+"""
+MCP Proxy Server (Proxy-Only Mode)
+
+This script provides a REST API with CORS for browser-based frontends
+to communicate with an already running Data Commons MCP server.
+
+Prerequisites:
+    Start the MCP server first:
+    python3 -m uv tool run datacommons-mcp serve http --port 3000
+
+Usage:
+    python main.py
+"""
+
+from src.config import _bootstrap_config_from_url
+from src.mcp.client import get_tools, initialize_mcp, MCP_PORT
+from src.server.app import app, PROXY_PORT
+import src.server.routes  # noqa: F401 (registers routes)
+
+
+def main():
+    """Main entry point."""
+    _bootstrap_config_from_url()
+
+    print("=" * 60)
+    print("Data Commons MCP Proxy Server (Proxy-Only Mode)")
+    print("=" * 60)
+    print(f"\nExpecting MCP server at: http://localhost:{MCP_PORT}")
+    print("\nMake sure you started the MCP server first:")
+    print(f"  python3 -m uv tool run datacommons-mcp serve http --port {MCP_PORT}")
+
+    # Try to connect to MCP server
+    print("\nChecking MCP server connection...")
+    if initialize_mcp():
+        tools = get_tools()
+        print(f"\nConnected! Found {len(tools)} tools:")
+        for t in tools:
+            print(f"  - {t.get('name')}")
+    else:
+        print("\nWARNING: Could not connect to MCP server")
+        print("The proxy will start anyway - MCP server can be started later")
+
+    # Start proxy
+    print(f"\nStarting proxy on port {PROXY_PORT}...")
+    print(f"Frontend should connect to: http://localhost:{PROXY_PORT}")
+    print("\nPress Ctrl+C to stop")
+    print("=" * 60)
+
+    app.run(host="0.0.0.0", port=PROXY_PORT, debug=False, threaded=True)
+
+
+if __name__ == "__main__":
+    main()
