+++ narratives/agent/src/mcp/client.py
@@ -1,0 +1,198 @@
+#!/usr/bin/env python3
+# Copyright 2024 Google LLC
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+import json
+import logging
+import os
+import time
+from typing import Any, Optional
+
+import requests
+
+from src.mcp.schema import fix_tool_arguments
+from src.session_logger import SessionLogger
+
+logger = logging.getLogger(__name__)
+
+# Configuration
+MCP_PORT = int(os.environ.get("MCP_PORT", 3000))
+MCP_URL = f"http://localhost:{MCP_PORT}/mcp"
+
+# Global state
+session_id = None
+tools_cache = None
+
+
+def mcp_request(method: str, params: dict = None, is_notification: bool = False) -> dict:
+    """Send a JSON-RPC request or notification to the MCP server.
+
+    Args:
+        method: The JSON-RPC method name
+        params: Optional parameters
+        is_notification: If True, sends as notification (no id, no response expected)
+    """
+    global session_id  # Needed to SET the global session_id from response headers
+
+    payload = {
+        "jsonrpc": "2.0",
+        "method": method
+    }
+
+    # Notifications don't have an id
+    if not is_notification:
+        payload["id"] = int(time.time() * 1000)
+
+    if params:
+        payload["params"] = params
+
+    headers = {
+        "Content-Type": "application/json",
+        "Accept": "application/json, text/event-stream"
+    }
+
+    if session_id:
+        headers["Mcp-Session-Id"] = session_id
+
+    try:
+        # For notifications, we send but don't expect a response
+        if is_notification:
+            requests.post(
+                MCP_URL,
+                json=payload,
+                headers=headers,
+                timeout=5
+            )
+            return {"result": "notification sent"}
+
+        response = requests.post(
+            MCP_URL,
+            json=payload,
+            headers=headers,
+            timeout=300,
+            stream=True
+        )
+
+        # Log response details for debugging
+        logger.info(f"MCP Response - Status: {response.status_code}, Headers: {dict(response.headers)}")
+
+        # Get session ID from response (try multiple header variations)
+        session_header = (
+            response.headers.get("Mcp-Session-Id") or
+            response.headers.get("mcp-session-id") or
+            response.headers.get("MCP-Session-ID")
+        )
+        if session_header:
+            session_id = session_header
+            logger.info(f"Got MCP session ID from headers: {session_id}")
+        else:
+            logger.warning(f"No session ID in response headers. Available headers: {list(response.headers.keys())}")
+
+        content_type = response.headers.get("content-type", "")
+
+        if "text/event-stream" in content_type:
+            # Parse SSE response
+            result = None
+            for line in response.iter_lines():
+                if line:
+                    line_str = line.decode('utf-8')
+                    if line_str.startswith("data: "):
+                        try:
+                            data = json.loads(line_str[6:])
+                            if "result" in data:
+                                result = data["result"]
+                            elif "error" in data:
+                                return {"error": data["error"]}
+                        except json.JSONDecodeError:
+                            continue
+            return {"result": result} if result else {"error": "No result"}
+        else:
+            return response.json()
+
+    except requests.exceptions.ConnectionError:
+        return {"error": f"Cannot connect to MCP server at {MCP_URL}. Make sure it's running!"}
+    except Exception as e:
+        return {"error": str(e)}
+
+
+def initialize_mcp() -> bool:
+    """Initialize the MCP session."""
+    global session_id
+
+    logger.info("Initializing MCP session...")
+
+    result = mcp_request("initialize", {
+        "protocolVersion": "2024-11-05",
+        "capabilities": {"roots": {"listChanged": True}},
+        "clientInfo": {"name": "dc-mcp-proxy", "version": "1.0.0"}
+    })
+
+    if "error" in result:
+        logger.error(f"Failed to initialize MCP: {result['error']}")
+        return False
+
+    logger.info(f"MCP session initialized: {session_id}")
+
+    # Send initialized notification (no id, no response expected)
+    mcp_request("notifications/initialized", {}, is_notification=True)
+    return True
+
+
+def get_tools() -> list:
+    """Get available tools from MCP server."""
+    global tools_cache
+
+    if tools_cache:
+        return tools_cache
+
+    result = mcp_request("tools/list", {})
+
+    if "result" in result and result["result"] and "tools" in result["result"]:
+        tools_cache = result["result"]["tools"]
+        return tools_cache
+
+    return []
+
+
+def call_tool(name: str, arguments: dict, session_logger: Optional[SessionLogger] = None) -> Any:
+    """Call a tool on the MCP server with optional logging."""
+    # Fix common parameter mistakes
+    fixed_args = fix_tool_arguments(name, arguments)
+    if fixed_args != arguments:
+        logger.info(f"Fixed arguments: {arguments} -> {fixed_args}")
+
+    # Log tool call request
+    if session_logger:
+        session_logger.log_mcp_tool_call(name, fixed_args)
+
+    start_time = time.time()
+
+    result = mcp_request("tools/call", {
+        "name": name,
+        "arguments": fixed_args
+    })
+
+    duration_ms = (time.time() - start_time) * 1000
+
+    if "result" in result:
+        # Log successful result
+        if session_logger:
+            session_logger.log_mcp_tool_result(name, result["result"], duration_ms, "success")
+        return result["result"]
+
+    # Log error result
+    error_result = {"error": result.get("error", "Unknown error")}
+    if session_logger:
+        session_logger.log_mcp_tool_result(name, error_result, duration_ms, "error")
+    return error_result
